chore: comprehensive codebase improvements and optimizations
- Refactored storage layer (LSM engine, blobstore, compaction)
- Enhanced SQL executor and query planner
- Updated database core CRUD operations and helpers
- Improved index implementations (column, spatial, text, timestamp, vector)
- Fixed transaction coordinator and WAL recovery
- Enhanced FFI and API interfaces
- Updated Vamana and DiskANN implementations
- Added merging iterator for efficient LSM operations
- Removed deprecated single_file_lsm.rs
- General code quality improvements and performance optimizations

+++ src/database/indexes/spatial.rs
@@ -71,7 +71,7 @@ impl MoteDB {
                     use std::hash::{Hash, Hasher};
                     let mut hasher = DefaultHasher::new();
                     table_name.hash(&mut hasher);
-                    let table_hash = (hasher.finish() & 0xFFFFFFFF) as u64;
+                    let table_hash = hasher.finish() & 0xFFFFFFFF;
                     
                     let start_key = table_hash << 32;
                     let end_key = (table_hash + 1) << 32;
