+++ src/database/indexes/vector.rs
@@ -65,7 +65,7 @@ impl MoteDB {
                     use std::hash::{Hash, Hasher};
                     let mut hasher = DefaultHasher::new();
                     table_name.hash(&mut hasher);
-                    let table_hash = (hasher.finish() & 0xFFFFFFFF) as u64;
+                    let table_hash = hasher.finish() & 0xFFFFFFFF;
                     
                     // composite_key格式: [table_hash:32位][row_id:32位]
                     let start_key = table_hash << 32;              // table的起始key
